+++ src/promptflow/tests/executor/unittests/processpool/test_line_execution_process_pool.py
@@ -1,4 +1,6 @@
 import uuid
+import os
+import multiprocessing
 from multiprocessing import Queue
 from pathlib import Path
 from tempfile import mkdtemp
@@ -11,7 +13,9 @@ from promptflow._utils.logger_utils import LogContext
 from promptflow.contracts.run_info import Status
 from promptflow.exceptions import ErrorTarget, UserErrorException
 from promptflow.executor import FlowExecutor
-from promptflow.executor._line_execution_process_pool import LineExecutionProcessPool, _exec_line
+from promptflow.executor._line_execution_process_pool import (
+    LineExecutionProcessPool, _exec_line, get_multiprocessing_context
+)
 from promptflow.executor.flow_executor import LineResult
 
 from ...utils import get_flow_sample_inputs, get_yaml_file
@@ -44,6 +48,25 @@ class TestLineExecutionProcessPool:
             else:
                 raise Exception(f"Invalid type of bulk input: {inputs}")
         return [self.get_line_inputs() for _ in range(nlinee)]
+
+    def create_line_execution_process_pool(self, dev_connections):
+        executor = FlowExecutor.create(
+            get_yaml_file(SAMPLE_FLOW),
+            dev_connections,
+            line_timeout_sec=1,
+        )
+        run_id = str(uuid.uuid4())
+        bulk_inputs = self.get_bulk_inputs()
+        nlines = len(bulk_inputs)
+        line_execution_process_pool = LineExecutionProcessPool(
+            executor,
+            nlines,
+            run_id,
+            "",
+            False,
+            None,
+        )
+        return line_execution_process_pool
 
     @pytest.mark.parametrize(
         "flow_folder",
@@ -167,3 +190,34 @@ class TestLineExecutionProcessPool:
             assert line_result.run_info.error["message"] == test_error_msg
             assert line_result.run_info.error["code"] == "UserError"
             assert line_result.run_info.status == Status.Failed
+
+    def test_process_set_environment_variable_successed(self, dev_connections):
+        os.environ["PF_BATCH_METHOD"] = "spawn"
+        line_execution_process_pool = self.create_line_execution_process_pool(dev_connections)
+        use_fork = line_execution_process_pool._use_fork
+        assert use_fork is False
+
+    def test_process_set_environment_variable_failed(self, dev_connections):
+        with patch("promptflow.executor._line_execution_process_pool.logger") as mock_logger:
+            mock_logger.warning.return_value = None
+            os.environ["PF_BATCH_METHOD"] = "test"
+            line_execution_process_pool = self.create_line_execution_process_pool(dev_connections)
+            use_fork = line_execution_process_pool._use_fork
+            assert use_fork == (multiprocessing.get_start_method() == "fork")
+            sys_start_methods = multiprocessing.get_all_start_methods()
+            exexpected_log_message = "Failed to set start method to 'test', start method test" \
+                f" is not in: {sys_start_methods}."
+            mock_logger.warning.assert_called_once_with(exexpected_log_message)
+
+    def test_process_not_set_environment_variable(self, dev_connections):
+        line_execution_process_pool = self.create_line_execution_process_pool(dev_connections)
+        use_fork = line_execution_process_pool._use_fork
+        assert use_fork == (multiprocessing.get_start_method() == "fork")
+
+    def test_get_multiprocessing_context(self):
+        # Set default start method to spawn
+        context = get_multiprocessing_context("spawn")
+        assert context.get_start_method() == "spawn"
+        # Not set start method
+        context = get_multiprocessing_context()
+        assert context.get_start_method() == multiprocessing.get_start_method()